feat(config): allow using the [name] placeholder in the name option
This allows use the same name as our Webpack output filename, but with the potential
to specify a fixed file extension. This is especially useful when concatenating multiple
types of files:

```
new ConcatTextPlugin({
  files: "res/*.{md,markdown}",
  name: "[name].md"
```

The concatenated markdown file will have the same name as , but since
we know that we're only collection markdown files, we can set the file extension explicitly here.